pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[project]
name = "taskmaster-pro"
version = "0.1.0"
description = "Task management app (FastAPI + Prisma + SQLite)"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "prisma>=0.15.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
  "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["."]
filterwarnings = [
  "ignore::DeprecationWarning",
]

[tool.uv]
dev-dependencies = []
